requirements.txt
1.scikit-learn
2.numpy
3.pandas
4.matplolib
